- moved components related to event tabs to `resouces/events/*` - components in `resources/events/*` are now acting as UI components 	- this will allow us to re-use these components later on
