+++ apps/webapp/src/components/resources/events/EventTabs/ActiveEvent/ActiveEventDetail.ts
@@ -2,14 +2,14 @@ import type { PropType } from 'vue'
 import { defineComponent, h } from 'vue'
 import type { EventCodec } from '@/services/api/resources/session/events/codec'
 import { EventTypesEnum } from '@/services/api/resources/session/events/constants'
-import ChangeUrlDetails from '@/components/pages/projects/ProjectSessionView/EventTabs/ActiveEvent/types/ChangeUrlDetails.vue'
-import DatabaseTransactionDetails from '@/components/pages/projects/ProjectSessionView/EventTabs/ActiveEvent/types/DatabaseTransactionDetails.vue'
-import ElementClickDetails from '@/components/pages/projects/ProjectSessionView/EventTabs/ActiveEvent/types/ElementClickDetails.vue'
-import NetworkEventDetails from '@/components/pages/projects/ProjectSessionView/EventTabs/ActiveEvent/types/NetworkEventDetails.vue'
-import ResizeScreenDetails from '@/components/pages/projects/ProjectSessionView/EventTabs/ActiveEvent/types/ResizeScreenDetails.vue'
-import ScrollTabDetails from '@/components/pages/projects/ProjectSessionView/EventTabs/ActiveEvent/types/ScrollTabDetails.vue'
-import LogEventDetails from '@/components/pages/projects/ProjectSessionView/EventTabs/ActiveEvent/types/LogEventDetails.vue'
-import CustomEvent from '@/components/pages/projects/ProjectSessionView/EventTabs/ActiveEvent/types/CustomEvent.vue'
+import ChangeUrlDetails from '@/components/resources/events/EventTabs/ActiveEvent/types/ChangeUrlDetails.vue'
+import DatabaseTransactionDetails from '@/components/resources/events/EventTabs/ActiveEvent/types/DatabaseTransactionDetails.vue'
+import ElementClickDetails from '@/components/resources/events/EventTabs/ActiveEvent/types/ElementClickDetails.vue'
+import NetworkEventDetails from '@/components/resources/events/EventTabs/ActiveEvent/types/NetworkEventDetails.vue'
+import ResizeScreenDetails from '@/components/resources/events/EventTabs/ActiveEvent/types/ResizeScreenDetails.vue'
+import ScrollTabDetails from '@/components/resources/events/EventTabs/ActiveEvent/types/ScrollTabDetails.vue'
+import LogEventDetails from '@/components/resources/events/EventTabs/ActiveEvent/types/LogEventDetails.vue'
+import CustomEvent from '@/components/resources/events/EventTabs/ActiveEvent/types/CustomEvent.vue'
 
 const ACTIVE_EVENT_DETAILS: Record<EventTypesEnum, ReturnType<typeof defineComponent>> = {
   [EventTypesEnum.CHANGED_URL]: ChangeUrlDetails,
