+++ apps/webapp/src/components/resources/events/EventTabs/ActiveEvent/BaseEventDetails.vue
