+++ apps/webapp/src/components/resources/events/EventTabs/ActiveEvent/common.ts
