+++ apps/webapp/src/components/resources/events/EventTabs/ActiveEvent/types/NetworkEventDetails.vue
@@ -130,8 +130,8 @@
 
 <script lang="ts" setup>
 import type { NetworkRequestEvent } from '@/services/api/resources/session/events/codec'
-import BaseEventDetails from '@/components/pages/projects/ProjectSessionView/EventTabs/ActiveEvent/BaseEventDetails.vue'
-import type { Props } from '@/components/pages/projects/ProjectSessionView/EventTabs/ActiveEvent/common'
+import BaseEventDetails from '@/components/resources/events/EventTabs/ActiveEvent/BaseEventDetails.vue'
+import type { Props } from '@/components/resources/events/EventTabs/ActiveEvent/common'
 import DCopyble from '@/components/DCopyble.vue'
 import { fromHeaderToText } from '@/services/resources/events/NetworkRequestEventService'
 import { useSessionsStore } from '@/stores/sessions'
