+++ apps/webapp/src/components/resources/events/EventTabs/ActiveEvent/types/ScrollTabDetails.vue
@@ -81,9 +81,9 @@
 </template>
 
 <script lang="ts" setup>
-import BaseEventDetails from '@/components/pages/projects/ProjectSessionView/EventTabs/ActiveEvent/BaseEventDetails.vue'
+import BaseEventDetails from '@/components/resources/events/EventTabs/ActiveEvent/BaseEventDetails.vue'
 import type { ScrollEvent } from '@/services/api/resources/session/events/codec'
-import type { Props } from '@/components/pages/projects/ProjectSessionView/EventTabs/ActiveEvent/common'
+import type { Props } from '@/components/resources/events/EventTabs/ActiveEvent/common'
 import DCopyble from '@/components/DCopyble.vue'
 
 const props = defineProps<Props<ScrollEvent>>()
