+++ apps/webapp/src/components/resources/events/EventTabs/types/ChangedUrlTab.vue
