+++ apps/webapp/src/components/resources/events/EventTabs/types/CustomEventTab.vue
