+++ apps/webapp/src/components/resources/events/EventTabs/types/DatabaseTransactionTab.vue
