+++ apps/webapp/src/components/resources/events/EventTabs/types/ElementClickTab.vue
