+++ apps/webapp/src/components/resources/events/EventTabs/types/LogEventTab.vue
