+++ apps/webapp/src/components/resources/events/EventTabs/types/NetworkEventTab.vue
