+++ apps/webapp/src/components/resources/events/EventTabs/types/ResizeScreenTab.vue
