+++ apps/webapp/src/views/projects/ProjectSessionView.vue
@@ -29,6 +29,8 @@
     <EventsSection
       data-cy="project-session-view__bottom-events-section"
       :key="sessionStore.activeSession.id"
+      :tabs="tabs"
+      :active-event-details="sessionStore.activeEventDetails"
       class="mt-2"
       v-if="isVideoConverted(sessionStore.activeSession.videoStatus)"
     />
@@ -45,16 +47,69 @@ import { isVideoConverted } from '@/services/resources/SessionsService'
 import { sessionsCodecFactory } from '@/services/factories/sessionsFactory'
 import { emptyPagination } from '@/services/api'
 import throttle from 'lodash.throttle'
-import { EventsSection } from '@/components/pages/projects/ProjectSessionView/EventTabs/EventsSection'
+import { EventsSection, TabsPropType } from '@/components/resources/events/EventTabs/EventsSection'
 import LivePreviewSection from '@/components/resources/session/LivePreviewSection.vue'
 import ActiveNetworkRequestResources from '@/components/pages/projects/ProjectSessionView/ActiveNetworkRequestResources.vue'
 import type { EventCodec } from '@/services/api/resources/session/events/codec'
+import { EventTypesEnum } from '@/services/api/resources/session/events/constants'
 
 const sessionStore = useSessionsStore()
 
 const isLoadingSession = ref(true)
 
 const route = useRoute()
+
+const tabs = computed<TabsPropType>(() => [
+  {
+    type: EventTypesEnum.NETWORK_REQUEST,
+    'data-cy': 'project-session-view__activate-network-events-tab',
+    icon: 'pi pi-desktop',
+    title: `Network (${sessionStore.networkEvents.length})`
+  },
+  {
+    type: EventTypesEnum.CHANGED_URL,
+    'data-cy': 'project-session-view__activate-change-url-events-tab',
+    icon: 'pi pi-at',
+    storeGetter: 'changedUrlEvents',
+    title: `URL (${sessionStore.changedUrlEvents.length})`
+  },
+  {
+    type: EventTypesEnum.ELEMENT_CLICKED,
+    'data-cy': 'project-session-view__activate-click-events-tab',
+    icon: 'pi pi-eye',
+    title: `Click (${sessionStore.clickEvents.length})`
+  },
+  {
+    type: EventTypesEnum.SCROLL,
+    'data-cy': 'project-session-view__activate-scroll-events-tab',
+    icon: 'pi pi-arrows-v',
+    title: `Scroll (${sessionStore.scrollEvents.length})`
+  },
+  {
+    type: EventTypesEnum.RESIZE_SCREEN,
+    'data-cy': 'project-session-view__activate-resize-events-tab',
+    icon: 'pi pi-arrows-h',
+    title: `Resize (${sessionStore.resizedScreenEvents.length})`
+  },
+  {
+    type: EventTypesEnum.DATABASE_TRANSACTION,
+    'data-cy': 'project-session-view__activate-db-transaction-events-tab',
+    icon: 'pi pi-server',
+    title: `Database (${sessionStore.databaseTransactionEvents.length})`
+  },
+  {
+    type: EventTypesEnum.LOG,
+    'data-cy': 'project-session-view__activate-log-events-tab',
+    icon: 'pi pi-paperclip',
+    title: `Logs (${sessionStore.logEvents.length})`
+  },
+  {
+    type: EventTypesEnum.CUSTOM_EVENT,
+    'data-cy': 'project-session-view__activate-custom-events-tab',
+    icon: 'pi pi-paperclip',
+    title: `Custom (${sessionStore.customEvents.length})`
+  }
+])
 
 onBeforeMount(async () => {
   isLoadingSession.value = true
